requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
appnope==0.1.4
astroid==3.3.9
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.21.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
dep-logic==0.4.11
dill==0.4.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dotenv==0.9.9
duckdb==1.3.1
dulwich==0.22.8
dydantic==0.0.8
email-validator==2.2.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flatbuffers==25.2.10
frozenlist==1.7.0
gcloud==0.18.3
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.39.0
google-cloud==0.34.0
google-genai==1.10.0
googleapis-common-protos==1.70.0
gql==3.5.3
graphql-core==3.2.6
greenlet==3.2.1
grpc-stubs==1.53.0.6
grpcio==1.73.1
grpcio-status==1.71.0
grpcio-tools==1.71.2
h11==0.16.0
hishel==0.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
id==1.5.0
idna==3.10
iniconfig==2.1.0
injector==0.21.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.35.0
isort==6.0.1
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.26
langchain-anthropic==0.3.18
langchain-aws==0.2.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.4
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.4.8
macholib==1.16.3
magika==0.6.1
markdown==3.8
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.6.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.3.1
oauth2client==4.1.3
onnxruntime==1.21.1
openai==1.97.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
openpyxl==3.1.5
orjson==3.11.0
ormsgpack==1.9.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
pbs-installer==2025.4.9
pdm==2.23.1
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.6.0
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydot==3.0.4
pygments==2.19.2
pygraph==0.3.0
pygraphviz==1.14
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyjwt==2.10.1
pylint==3.3.6
pyparsing==3.2.3
pyproject-hooks==1.2.0
pytest==8.3.5
python-call-graph==2.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
resolvelib==1.1.0
rich==12.6.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.39
truststore==0.10.1
typer==0.9.4
types-protobuf==5.29.1.20250403
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unearth==0.17.5
urllib3==2.5.0
uuid==1.30
uv==0.6.14
uvicorn==0.23.2
virtualenv==20.30.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
